requirements.txt
# Dependências principais
playwright==1.42.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9

# Dependências de desenvolvimento (opcional)
pytest==8.0.2
pytest-playwright==0.4.0

# Dependências adicionais recomendadas
pandas==2.2.1  # Para manipulação de dados
openpyxl==3.1.2  # Se precisar exportar para Excel